# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CheckoutInput {
  email: string;
  password: string;
  amount: number | string;
}

export interface CheckoutResponse {
  order_id: string;
  status: string;
  pix_copy_paste: string;
  pix_qr_code: string;
  expires_at?: string;
  amount?: number;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface UserProfile {
  id: number;
  email: string;
  role: string;
}

export interface AuthSession {
  token: string;
  user: UserProfile;
}

export interface AccessItem {
  id: number;
  externalOrderId: string;
  grantedAt: string;
}

export interface AccessResponse {
  hasAccess: boolean;
  items: AccessItem[];
}

export interface NexusWebhookPayload {
  id?: string;
  status?: string;
  email?: string;
  amount?: number;
  created_at?: string;
  [key: string]: unknown;
 }

export interface WebhookAck {
  received: boolean;
  processed: boolean;
}

export type CreateContentInputType = typeof CreateContentInputType[keyof typeof CreateContentInputType];


export const CreateContentInputType = {
  album: 'album',
  video: 'video',
} as const;

export interface CreateContentInput {
  title: string;
  description?: string;
  type?: CreateContentInputType;
  price?: number;
  teaserUrl?: string;
  privateFolderKey?: string;
}

export interface ContentRecord {
  id: number;
  title: string;
  description?: string | null;
  type: string;
  price?: number | null;
  teaserUrl?: string | null;
  privateFolderKey?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ContentPublic {
  id: number;
  title: string;
  description?: string | null;
  type: string;
  price?: number | null;
  teaserUrl?: string | null;
  createdAt: string;
}

export interface SecureStreamResponse {
  contentId: number;
  streamUrl?: string | null;
  expiresAt: string;
  signature: string;
}

export interface OrderRecord {
  id: number;
  externalId: string;
  email: string;
  status: string;
  amount?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface UserRecord {
  id: number;
  email: string;
  role: string;
  createdAt: string;
}

export interface AdminStats {
  totalOrders: number;
  paidOrders: number;
  totalRevenue: number;
  totalUsers: number;
  totalContents: number;
  totalAccesses: number;
}

export interface ApiError {
  error: string;
}

export type NexusWebhookParams = {
token?: string;
};