# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: payments
    description: Payment and checkout operations
  - name: auth
    description: Authentication operations
  - name: members
    description: Member area operations
  - name: contents
    description: Content management
  - name: admin
    description: Admin-only operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Create a Pix checkout via Nexus Pag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Pix order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "502":
          description: Gateway error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /me/access:
    get:
      operationId: getMyAccess
      tags: [members]
      summary: Get authenticated user access list
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Access list returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /payments/nexus-webhook:
    post:
      operationId: nexusWebhook
      tags: [payments]
      summary: Nexus Pag payment webhook
      parameters:
        - name: token
          in: query
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NexusWebhookPayload"
      responses:
        "200":
          description: Webhook received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookAck"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /contents:
    get:
      operationId: listContents
      tags: [contents]
      summary: List all available contents (public metadata)
      responses:
        "200":
          description: Contents list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentPublic"

  /contents/{id}/secure-stream:
    get:
      operationId: getSecureStream
      tags: [contents]
      summary: Get a time-limited secure stream URL for a content item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Secure stream URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SecureStreamResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden — no active access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Content not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin/contents:
    post:
      operationId: createContent
      tags: [admin]
      summary: Create a new content entry (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContentInput"
      responses:
        "201":
          description: Content created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentRecord"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin/users:
    get:
      operationId: listAdminUsers
      tags: [admin]
      summary: List all users (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserRecord"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get dashboard summary stats (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    CheckoutInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        amount:
          oneOf:
            - type: number
            - type: string
      required: [email, password, amount]

    CheckoutResponse:
      type: object
      properties:
        order_id:
          type: string
        status:
          type: string
        pix_copy_paste:
          type: string
        pix_qr_code:
          type: string
        expires_at:
          type: string
        amount:
          type: number
      required: [order_id, status, pix_copy_paste, pix_qr_code]

    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]

    AuthSession:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [token, user]

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
      required: [id, email, role]

    AccessResponse:
      type: object
      properties:
        hasAccess:
          type: boolean
        items:
          type: array
          items:
            $ref: "#/components/schemas/AccessItem"
      required: [hasAccess, items]

    AccessItem:
      type: object
      properties:
        id:
          type: integer
        externalOrderId:
          type: string
        grantedAt:
          type: string
      required: [id, externalOrderId, grantedAt]

    NexusWebhookPayload:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        email:
          type: string
        amount:
          type: number
        created_at:
          type: string
      additionalProperties: true

    WebhookAck:
      type: object
      properties:
        received:
          type: boolean
        processed:
          type: boolean
      required: [received, processed]

    CreateContentInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [album, video]
        price:
          type: number
        teaserUrl:
          type: string
        privateFolderKey:
          type: string
      required: [title]

    ContentRecord:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
        price:
          type: integer
          nullable: true
        teaserUrl:
          type: string
          nullable: true
        privateFolderKey:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, type, createdAt, updatedAt]

    ContentPublic:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        type:
          type: string
        price:
          type: integer
          nullable: true
        teaserUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, title, type, createdAt]

    SecureStreamResponse:
      type: object
      properties:
        contentId:
          type: integer
        streamUrl:
          type: string
          nullable: true
        expiresAt:
          type: string
        signature:
          type: string
      required: [contentId, expiresAt, signature]

    OrderRecord:
      type: object
      properties:
        id:
          type: integer
        externalId:
          type: string
        email:
          type: string
        status:
          type: string
        amount:
          type: integer
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, externalId, email, status, createdAt, updatedAt]

    UserRecord:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        role:
          type: string
        createdAt:
          type: string
      required: [id, email, role, createdAt]

    AdminStats:
      type: object
      properties:
        totalOrders:
          type: integer
        paidOrders:
          type: integer
        totalRevenue:
          type: integer
        totalUsers:
          type: integer
        totalContents:
          type: integer
        totalAccesses:
          type: integer
      required: [totalOrders, paidOrders, totalRevenue, totalUsers, totalContents, totalAccesses]

    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]