# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalOrders: number;
  paidOrders: number;
  totalRevenue: number;
  totalUsers: number;
  totalContents: number;
  totalAccesses: number;
}