# lib/api-zod/src/generated/types/contentPublic.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentPublic {
  id: number;
  title: string;
  description?: string | null;
  type: string;
  price?: number | null;
  teaserUrl?: string | null;
  createdAt: string;
}