# lib/api-zod/src/generated/types/contentRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContentRecord {
  id: number;
  title: string;
  description?: string | null;
  type: string;
  price?: number | null;
  teaserUrl?: string | null;
  privateFolderKey?: string | null;
  createdAt: string;
  updatedAt: string;
}