# lib/api-zod/src/generated/types/createContentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContentInputType } from './createContentInputType';

export interface CreateContentInput {
  title: string;
  description?: string;
  type?: CreateContentInputType;
  price?: number;
  teaserUrl?: string;
  privateFolderKey?: string;
}