# lib/api-zod/src/generated/types/orderRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OrderRecord {
  id: number;
  externalId: string;
  email: string;
  status: string;
  amount?: number | null;
  createdAt: string;
  updatedAt: string;
}